requirements.txt
spidev
rpi.gpio
gpiozero
